# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blog API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password",
                  "confirmPassword"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "newuser",
                    "minLength": 1,
                    "maxLength": 50,
                    "pattern": "^[a-zA-Z0-9._-]+$",
                    "description": "Must contain only letters, numbers, dots, hyphens, or underscores."
                  },
                  "password": {
                    "type": "string",
                    "example": "$Passw0rd1111",
                    "minLength": 8,
                    "maxLength": 50,
                    "description": "Must contain at least 1 uppercase letter, 1 lowercase letter, 1 number, and 1 special character."
                  },
                  "avatarUrl": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://example.com/avatar.jpg",
                    "minLength": 15,
                    "maxLength": 200
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "headers": {
              "Location": {
                "description": "URL of the created user profile",
                "schema": {
                  "type": "string",
                  "example": "/users/user1"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "example": "aaa.bbb"
                    },
                    "refreshToken": {
                      "type": "string",
                      "example": "bbb.ccc"
                    },
                    "id": {
                      "type": "string",
                      "example": "user1"
                    },
                    "username": {
                      "type": "string",
                      "example": "johndoe"
                    },
                    "avatarUrl": {
                      "type": "string",
                      "format": "uri",
                      "example": "https://example.com/avatar.jpg"
                    },
                    "role": {
                      "type": "string",
                      "example": "user"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login with username and password",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "newuser",
                    "minLength": 1,
                    "maxLength": 50,
                    "pattern": "^[a-zA-Z0-9._-]+$",
                    "description": "Must contain only letters, numbers, dots, hyphens, or underscores."
                  },
                  "password": {
                    "type": "string",
                    "example": "$Passw0rd1111",
                    "minLength": 8,
                    "maxLength": 50,
                    "description": "Must match the password field."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "example": "aaa.bbb"
                    },
                    "refreshToken": {
                      "type": "string",
                      "example": "bbb.ccc"
                    },
                    "id": {
                      "type": "string",
                      "example": "user1"
                    },
                    "username": {
                      "type": "string",
                      "example": "johndoe"
                    },
                    "avatarUrl": {
                      "type": "string",
                      "format": "uri",
                      "example": "https://example.com/avatar.jpg"
                    },
                    "role": {
                      "type": "string",
                      "example": "user"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/auth/change-password": {
      "post": {
        "summary": "Change current user's password",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "currentPassword",
                  "newPassword",
                  "confirmPassword"
                ],
                "properties": {
                  "currentPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string",
                    "example": "$Passw0rd1111",
                    "minLength": 8,
                    "maxLength": 50,
                    "description": "Must match the password field."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "example": "aaa.bbb"
                    },
                    "refreshToken": {
                      "type": "string",
                      "example": "bbb.ccc"
                    },
                    "id": {
                      "type": "string",
                      "example": "user1"
                    },
                    "username": {
                      "type": "string",
                      "example": "johndoe"
                    },
                    "avatarUrl": {
                      "type": "string",
                      "format": "uri",
                      "example": "https://example.com/avatar.jpg"
                    },
                    "role": {
                      "type": "string",
                      "example": "user"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized or current password incorrect"
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "Refresh access token",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "example": "aaaaaa.bbbbbb.cccccc"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New access token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "example": "aaa.bbb"
                    },
                    "refreshToken": {
                      "type": "string",
                      "example": "bbb.ccc"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Invalid or expired refresh token"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Logout current user",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Logout successful"
          }
        }
      }
    },
    "/api/auth/oauth/{provider}": {
      "get": {
        "summary": "Start OAuth flow with a provider",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "google",
                "github"
              ]
            },
            "description": "OAuth provider name"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirects to the provider's OAuth consent screen"
          },
          "400": {
            "description": "Invalid provider"
          }
        }
      }
    },
    "/api/auth/oauth/{provider}/callback": {
      "get": {
        "summary": "OAuth callback",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "google",
                "github"
              ]
            },
            "description": "OAuth provider name"
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth login success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "example": "aaa.bbb"
                    },
                    "refreshToken": {
                      "type": "string",
                      "example": "bbb.ccc"
                    },
                    "id": {
                      "type": "string",
                      "example": "user1"
                    },
                    "username": {
                      "type": "string",
                      "example": "johndoe"
                    },
                    "avatarUrl": {
                      "type": "string",
                      "format": "uri",
                      "example": "https://example.com/avatar.jpg"
                    },
                    "role": {
                      "type": "string",
                      "example": "user"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "OAuth failed"
          }
        }
      }
    },
    "/api/auth/oauth/unlink/{provider}": {
      "delete": {
        "summary": "Unlink an OAuth provider",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "google",
                "github"
              ]
            },
            "description": "OAuth provider to unlink"
          }
        ],
        "responses": {
          "200": {
            "description": "Unlinked successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}